Split parser tests into general/bcr command
