+++ Unit4.Automation.Tests/Unit4.Automation.Tests.csproj
@@ -93,6 +93,7 @@
     <Compile Include="CostCentreHierarchyTests.cs" />
     <Compile Include="BcrReportTests.cs" />
     <Compile Include="ResqlTests.cs" />
+    <Compile Include="Parser\BcrCommandParserTests.cs" />
     <Compile Include="Parser\CommandParserTests.cs" />
     <Compile Include="SerializableRoundTripTests.cs" />
     <Compile Include="CacheTests.cs" />
